apps/backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "5data-hrms"
version = "1.0.0"
description = "Production-grade Enterprise HRMS & Timesheet System"
requires-python = ">=3.11"
dependencies = [
    "Django==4.2.8",
    "djangorestframework==3.14.0",
    "django-cors-headers==4.3.1",
    "django-environ==0.21.0",
    "djangorestframework-simplejwt==5.3.2",
    "Pillow==10.1.0",
    "psycopg2-binary==2.9.9",
    "python-decouple==3.8",
    "gunicorn==21.2.0",
    "celery==5.3.4",
    "redis==5.0.1",
]

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "flake8==6.1.0",
    "isort==5.13.2",
    "pytest==7.4.3",
    "pytest-django==4.7.0",
    "pytest-cov==4.1.0",
]

